HistoryTokenSaveValueAnchorComponent & ValueHistoryTokenAnchorComponent moved to anchor package
