+++ src/main/java/walkingkooka/spreadsheet/dominokit/reference/SpreadsheetSelectionDeleteAnchorComponent.java
@@ -17,14 +17,14 @@
 
 package walkingkooka.spreadsheet.dominokit.reference;
 
+import walkingkooka.spreadsheet.dominokit.anchor.ValueHistoryTokenAnchorComponent;
+import walkingkooka.spreadsheet.dominokit.anchor.ValueHistoryTokenAnchorComponentDelegator;
 import walkingkooka.spreadsheet.dominokit.history.HistoryContext;
 import walkingkooka.spreadsheet.dominokit.history.HistoryToken;
 import walkingkooka.spreadsheet.dominokit.history.HistoryTokenAnchorComponent;
 import walkingkooka.spreadsheet.dominokit.history.SpreadsheetCellDeleteHistoryToken;
 import walkingkooka.spreadsheet.dominokit.history.SpreadsheetColumnDeleteHistoryToken;
 import walkingkooka.spreadsheet.dominokit.history.SpreadsheetRowDeleteHistoryToken;
-import walkingkooka.spreadsheet.dominokit.value.ValueHistoryTokenAnchorComponent;
-import walkingkooka.spreadsheet.dominokit.value.ValueHistoryTokenAnchorComponentDelegator;
 import walkingkooka.spreadsheet.reference.SpreadsheetSelection;
 
 import java.util.Objects;
